Append blog Site Title to dynamic output
Since this is a dynamic block, we can modify the markup at run time
(instead of saving the markup in the database to be displayed).

In this case, we are appending the Site Title of the blog to the message being
output.

If we were to go to
Settings > General
and modify the "Site Title" setting, the rendered output would change
(without our re-opening the post in Gutenberg).

See #203